As per the man page of hwclock, section 'Automatic Hardware Clock Synchronization by the Kernel': - the system clock is in UTC (by default) - the '11 minute mode' is turned on by ntpd
Setting the hardware clock in UTC makes things easier as the system clock will be
initialised at boot time from the hardware clock with no need for in-kernel timezone.
And the hardware clock will synchronised every 11 minutes from the system clock, itself
synchronised from the ntp daemon.

So, the calls to hwclock in the initramfs/initscript are not needed.
Just ensure that the hardware clock is in UTC mode and enable the ntp daemon.